+++ MercadoPagoSDK/MercadoPagoSDK/Models/OneTap/PXOneTapDto.swift
@@ -61,32 +61,3 @@ open class PXOneTapDto: NSObject, Codable {
         return try JSONDecoder().decode(PXOneTapDto.self, from: data)
     }
 }
-extension PXOneTapDto {
-    func getAccountMoneyForTracking() -> [String: Any] {
-        var accountMoneyDic: [String: Any] = [:]
-        accountMoneyDic["payment_method_type"] = paymentTypeId
-        accountMoneyDic["payment_method_id"] = paymentMethodId
-        var extraInfo: [String: Any] = [:]
-        extraInfo["balance"] = accountMoney?.availableBalance
-        extraInfo["invested"] = accountMoney?.invested
-        accountMoneyDic["extra_info"] = extraInfo
-
-        return accountMoneyDic
-    }
-
-    func getCardForTracking() -> [String: Any] {
-        var savedCardDic: [String: Any] = [:]
-        savedCardDic["payment_method_type"] = paymentTypeId
-        savedCardDic["payment_method_id"] = paymentMethodId
-        var extraInfo: [String: Any] = [:]
-        extraInfo["card_id"] = oneTapCard?.cardId
-        let cardIdsEsc = PXTrackingStore.sharedInstance.getData(forKey: PXTrackingStore.cardIdsESC) as? [String] ?? []
-        extraInfo["has_esc"] = cardIdsEsc.contains(oneTapCard?.cardId ?? "")
-        extraInfo["selected_installment"] = oneTapCard?.selectedPayerCost?.getPayerCostForTracking()
-        if let issuerId = oneTapCard?.cardUI?.issuerId {
-            extraInfo["issuer_id"] = Int64(issuerId)
-        }
-        savedCardDic["extra_info"] = extraInfo
-        return savedCardDic
-    }
-}
